requirements.txt
telethon==1.28.5
python-dotenv==1.0.0
